requirements.txt
streamlit
langchain
langchain-groq
python-docx
pandas
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
